+++ Move.py
@@ -1,0 +1,5 @@
+class Move:
+	def __init__(self, dest_rank, from_rank, stack):
+		self.dest_rank = dest_rank
+		self.from_rank = from_rank
+		self.stack = stack
